requirements.txt
arrow==1.2.3
asgiref==3.6.0
beautifulsoup4==4.11.1
binaryornot==0.4.4
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
cookiecutter==2.1.1
Django==4.1.5
django-bootstrap-v5==1.0.11
django-braces==1.15.0
django-environ==0.9.0
django-extensions==3.2.1
django_debug_toolbar==3.8.1
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.2
misaka==2.1.1
Pillow==9.4.0
psycopg2-binary==2.9.5
pycparser==2.21
python-dateutil==2.8.2
python-slugify==7.0.0
PyYAML==6.0
requests==2.28.2
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
text-unidecode==1.3
urllib3==1.26.14
